+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @react-elf/docs
 
+## 0.0.71
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @react-elf/ui@0.0.71
+
+## 0.0.70
+
+### Patch Changes
+
+- add Alert mdx
+
 ## 0.0.69
 
 ### Patch Changes