+++ packages/@react-elf/alert/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @react-elf/button
 
+## 0.0.71
+
+### Patch Changes
+
+- modify alert component
+
+- Updated dependencies []:
+  - @react-elf/shared@0.0.71
+
 ## 0.0.69
 
 ### Patch Changes