+++ packages/@react-elf/popover/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @react-elf/popover
 
+## 0.0.71
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @react-elf/shared@0.0.71
+
 ## 0.0.69
 
 ### Patch Changes